requirements.txt
Pillow>=9.5.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
qasync>=0.24.0
numpy>=1.24.3